# Optimize `partial!` DSL