--- a/Gemfile
+++ b/Gemfile
@@ -5,3 +5,5 @@ gemspec
 gem "rake"
 gem "mocha", require: false
 gem "appraisal"
+gem 'benchmark-ips', require: false
+gem 'benchmark-memory', require: false