--- a/benchmarks/partials/partial_dsl.rb
+++ b/benchmarks/partials/partial_dsl.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'benchmark/ips'
 require 'benchmark/memory'
 require_relative 'setup'