--- a/lib/jbuilder/jbuilder_template.rb
+++ b/lib/jbuilder/jbuilder_template.rb
@@ -49,11 +49,12 @@ class JbuilderTemplate < Jbuilder
   #
   #   json.comments @post.comments, partial: "comments/comment", as: :comment, cached: true
   #
-  def partial!(*args)
-    if args.one? && _is_active_model?(args.first)
-      _render_active_model_partial args.first
+  def partial!(partial_or_model = nil, **options)
+    if partial_or_model && _is_active_model?(partial_or_model)
+      _render_active_model_partial partial_or_model
     else
-      _render_explicit_partial(*args)
+      options[:partial] = partial_or_model if partial_or_model
+      _render_partial_with_options options
     end
   end
 
@@ -120,7 +121,7 @@ class JbuilderTemplate < Jbuilder
 
     if args.one? && _partial_options?(options)
       options[:collection] = collection
-      partial! options
+      _render_partial_with_options options
     else
       super
     end
@@ -243,28 +244,6 @@ class JbuilderTemplate < Jbuilder
     end
 
     _set_value name, value
-  end
-
-  def _render_explicit_partial(name_or_options, locals = {})
-    case name_or_options
-    when ::Hash
-      # partial! partial: 'name', foo: 'bar'
-      options = name_or_options
-    else
-      # partial! 'name', locals: {foo: 'bar'}
-      if locals.one? && (locals.keys.first == :locals)
-        locals[:partial] = name_or_options
-        options = locals
-      else
-        options = { partial: name_or_options, locals: locals }
-      end
-      # partial! 'name', foo: 'bar'
-      as = locals.delete(:as)
-      options[:as] = as if as.present?
-      options[:collection] = locals[:collection] if locals.key?(:collection)
-    end
-
-    _render_partial_with_options options
   end
 
   def _render_active_model_partial(object)