Addressing some hotspots we're seeing in our profiling. Instead of doing a full context dump here, I'll annotate the points of interest below.

Note on the benchmarks:
- When it comes to running `jbuilder` that need partials, a lot more needs to be setup with `actionview`. This results in _more stuff_ running under the benchmark, which dilutes the measurement. I did my best to come get a minimal setup. I'm sure there are ways to lean it out more, but it's a lot to figure out.
- The IPS benchmarks might not look super convincing, but these _do_ come up as hotspots in our profiling. `partial!` can run many, many times per template render, and it's hard to capture this in IPS because the more complex the template is, the fewer times it can run per second, and thus the sample size becomes much smaller.
- For the memory benchmark, I am running the DSL 100 times. I picked that number somewhat arbitrarily, but I want to run it more than once to make the memory optimizations clearer. Something like `partial!` will often run more than once in a given a template, so seeing something that reflects that helps highlight the savings we hope to see.

Compared to the upstream `jbuilder` (ie. the `upstream_main` branch). This compares all of our optimizations done so far against "stock" `jbuilder`.

```
ruby 3.4.4 (2025-05-14 revision a38531fd3f) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
              before     6.437k i/100ms
               after     6.917k i/100ms
Calculating -------------------------------------
              before     70.077k (± 3.6%) i/s   (14.27 μs/i) -    354.035k in   5.058645s
               after     76.812k (± 5.3%) i/s   (13.02 μs/i) -    387.352k in   5.057574s

Comparison:
               after:    76812.1 i/s
              before:    70076.6 i/s - 1.10x  slower
```

```
Calculating -------------------------------------
              before   419.200k memsize (     0.000  retained)
                         4.800k objects (     0.000  retained)
                        11.000  strings (     0.000  retained)
               after   303.200k memsize (     0.000  retained)
                         3.700k objects (     0.000  retained)
                         9.000  strings (     0.000  retained)

Comparison:
               after:     303200 allocated
              before:     419200 allocated - 1.38x more
```

Compared to `main`, which includes all of our other optimizations already. This more closely reflects the impact this optimization has compared to what we're currently running.

```
ruby 3.4.4 (2025-05-14 revision a38531fd3f) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
              before     6.991k i/100ms
               after     7.327k i/100ms
Calculating -------------------------------------
              before     70.618k (± 4.2%) i/s   (14.16 μs/i) -    356.541k in   5.058201s
               after     75.337k (± 4.2%) i/s   (13.27 μs/i) -    381.004k in   5.065986s

Comparison:
               after:    75336.8 i/s
              before:    70618.2 i/s - same-ish: difference falls within error
```

```
Calculating -------------------------------------
              before   315.200k memsize (     0.000  retained)
                         4.000k objects (     0.000  retained)
                         9.000  strings (     0.000  retained)
               after   303.200k memsize (     0.000  retained)
                         3.700k objects (     0.000  retained)
                         9.000  strings (     0.000  retained)

Comparison:
               after:     303200 allocated
              before:     315200 allocated - 1.04x more
```

(I will include the benchmarks in this PR soon-ish... just want to clean it up first so that it doesn't look like nonsense.)